email_ingest_starter_with_orders_api_ui_labels/email_ingest_starter/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
alembic==1.13.2
pydantic==2.9.2
cryptography==43.0.1
python-multipart==0.0.9
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1

APScheduler==3.10.4
python-dateutil==2.9.0.post0
